requirements.txt
hydra-core >=1.2
selenium==4.2.0
yfinance>=0.1.7.0
omegaconf~=2.2.2
setuptools==62.3.2